call boost::posix_time::milliseconds with integral parameter
With Boost >= 1.67 the constructor of various boost::posix_time classes
requires the initialisation parameters to be integral. This patch makes
sure an integral value is passed, instead of relying on implicit
conversions.

+++ include/actionlib/server/simple_action_server_imp.h
@@ -37,6 +37,7 @@
 #ifndef ACTIONLIB__SERVER__SIMPLE_ACTION_SERVER_IMP_H_
 #define ACTIONLIB__SERVER__SIMPLE_ACTION_SERVER_IMP_H_
 
+#include <cstdint>
 #include <string>
 
 namespace actionlib
@@ -394,7 +395,7 @@ void SimpleActionServer<ActionSpec>::executeLoop()
       }
     } else {
       execute_condition_.timed_wait(lock,
-        boost::posix_time::milliseconds(loop_duration.toSec() * 1000.0f));
+        boost::posix_time::milliseconds((int64_t)(loop_duration.toSec() * 1000.0f)));
     }
   }
 }
